Improve Android testing framework.

+++ core_app/src/main/java/org/opendatakit/common/android/utilities/ODKDatabaseImplUtils.java
@@ -15,16 +15,12 @@
  */
 package org.opendatakit.common.android.utilities;
 
-import java.io.File;
-import java.io.IOException;
-import java.util.ArrayList;
-import java.util.Collection;
-import java.util.Collections;
-import java.util.HashMap;
-import java.util.List;
-import java.util.Map;
-import java.util.TreeMap;
-import java.util.UUID;
+import android.content.ContentValues;
+import android.database.Cursor;
+
+import com.fasterxml.jackson.core.JsonParseException;
+import com.fasterxml.jackson.core.JsonProcessingException;
+import com.fasterxml.jackson.databind.JsonMappingException;
 
 import org.apache.commons.io.FileUtils;
 import org.apache.commons.io.filefilter.IOFileFilter;
@@ -39,9 +35,9 @@ import org.opendatakit.aggregate.odktables.rest.entity.Column;
 import org.opendatakit.common.android.data.ColorRule;
 import org.opendatakit.common.android.data.ColumnDefinition;
 import org.opendatakit.common.android.data.OrderedColumns;
+import org.opendatakit.common.android.data.Row;
 import org.opendatakit.common.android.data.TableDefinitionEntry;
 import org.opendatakit.common.android.data.UserTable;
-import org.opendatakit.common.android.data.Row;
 import org.opendatakit.common.android.database.DatabaseConstants;
 import org.opendatakit.common.android.database.DatabaseFactory;
 import org.opendatakit.common.android.database.OdkDatabase;
@@ -54,12 +50,16 @@ import org.opendatakit.common.android.utilities.StaticStateManipulator.IStaticFi
 import org.opendatakit.database.service.KeyValueStoreEntry;
 import org.sqlite.database.sqlite.SQLiteException;
 
-import android.content.ContentValues;
-import android.database.Cursor;
-
-import com.fasterxml.jackson.core.JsonParseException;
-import com.fasterxml.jackson.core.JsonProcessingException;
-import com.fasterxml.jackson.databind.JsonMappingException;
+import java.io.File;
+import java.io.IOException;
+import java.util.ArrayList;
+import java.util.Collection;
+import java.util.Collections;
+import java.util.HashMap;
+import java.util.List;
+import java.util.Map;
+import java.util.TreeMap;
+import java.util.UUID;
 
 public class ODKDatabaseImplUtils {
 
@@ -187,8 +187,6 @@ public class ODKDatabaseImplUtils {
    * Perform a query with the given parameters.
    * 
    * @param db
-   * @param distinct
-   *          true if each returned row should be distinct (collapse duplicates)
    * @param table
    * @param columns
    * @param selection
@@ -203,7 +201,7 @@ public class ODKDatabaseImplUtils {
       String[] selectionArgs, String groupBy, String having, String orderBy, String limit)
       throws SQLiteException {
     Cursor c = db.queryDistinct(table, columns, selection, selectionArgs, groupBy, having, orderBy,
-        limit);
+            limit);
     return c;
   }
 
@@ -366,7 +364,7 @@ public class ODKDatabaseImplUtils {
       OrderedColumns orderedDefns, String rowId) {
 
     UserTable table = rawSqlQuery(db, appName, tableId, orderedDefns, DataTableColumns.ID + "=?",
-        new String[] { rowId }, null, null, DataTableColumns.SAVEPOINT_TIMESTAMP, "DESC");
+            new String[]{rowId}, null, null, DataTableColumns.SAVEPOINT_TIMESTAMP, "DESC");
 
     return table;
   }
@@ -604,41 +602,44 @@ public class ODKDatabaseImplUtils {
 
     String assetsCsvDir = ODKFileUtils.getAssetsCsvFolder(appName);
     try {
-      Collection<File> files = FileUtils.listFiles(new File(assetsCsvDir), new IOFileFilter() {
+      File file = new File(assetsCsvDir);
+      if(file.exists()) {
+        Collection<File> files = FileUtils.listFiles(file, new IOFileFilter() {
 
-        @Override
-        public boolean accept(File file) {
-          String[] parts = file.getName().split("\\.");
-          return (parts[0].equals(tableId) && parts[parts.length - 1].equals("csv") && (parts.length == 2
-              || parts.length == 3 || (parts.length == 4 && parts[parts.length - 2]
-              .equals("properties"))));
+          @Override
+          public boolean accept(File file) {
+            String[] parts = file.getName().split("\\.");
+            return (parts[0].equals(tableId) && parts[parts.length - 1].equals("csv") && (parts.length == 2
+                    || parts.length == 3 || (parts.length == 4 && parts[parts.length - 2]
+                    .equals("properties"))));
+          }
+
+          @Override
+          public boolean accept(File dir, String name) {
+            String[] parts = name.split("\\.");
+            return (parts[0].equals(tableId) && parts[parts.length - 1].equals("csv") && (parts.length == 2
+                    || parts.length == 3 || (parts.length == 4 && parts[parts.length - 2]
+                    .equals("properties"))));
+          }
+        }, new IOFileFilter() {
+
+          // don't traverse into directories
+          @Override
+          public boolean accept(File arg0) {
+            return false;
+          }
+
+          // don't traverse into directories
+          @Override
+          public boolean accept(File arg0, String arg1) {
+            return false;
+          }
+        });
+
+        FileUtils.deleteDirectory(new File(tableDir));
+        for (File f : files) {
+          FileUtils.deleteQuietly(f);
         }
-
-        @Override
-        public boolean accept(File dir, String name) {
-          String[] parts = name.split("\\.");
-          return (parts[0].equals(tableId) && parts[parts.length - 1].equals("csv") && (parts.length == 2
-              || parts.length == 3 || (parts.length == 4 && parts[parts.length - 2]
-              .equals("properties"))));
-        }
-      }, new IOFileFilter() {
-
-        // don't traverse into directories
-        @Override
-        public boolean accept(File arg0) {
-          return false;
-        }
-
-        // don't traverse into directories
-        @Override
-        public boolean accept(File arg0, String arg1) {
-          return false;
-        }
-      });
-
-      FileUtils.deleteDirectory(new File(tableDir));
-      for (File f : files) {
-        FileUtils.deleteQuietly(f);
       }
     } catch (IOException e1) {
       e1.printStackTrace();
